Refactored Retries util to not use Java 8 Predicate.
	new file:   xchange-core/src/main/java/com/xeiam/xchange/utils/retries/IPredicate.java
	modified:   xchange-core/src/main/java/com/xeiam/xchange/utils/retries/Retries.java
	modified:   xchange-examples/src/main/java/com/xeiam/xchange/examples/core/utils/RetriesDemo.java
