+++ xchange-core/src/main/java/com/xeiam/xchange/utils/retries/IPredicate.java
@@ -1,0 +1,5 @@
+package com.xeiam.xchange.utils.retries;
+
+public interface IPredicate<T> {
+	public boolean test(T t);
+}
